+++ src/com/aleexf/game/world/cell/Bomb.kt
@@ -1,3 +1,3 @@
 package com.aleexf.game.world.cell
 
-class Bomb(x:Int, y:Int, val delay:Int):Object(x, y, true, 3)
+class Bomb(x:Int, y:Int, val delay:Long):Object(x, y, true, 3)
